TP1_material/python/requirements.txt
PyOpenGL
glfw
numpy
PyGLM